Pull image subclasses into their own files
